requirements.txt
instagrapi==2.1.3
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.3
python-dateutil==2.8.2
APScheduler==3.10.4
random-username==1.0.2
faker==24.2.0
jinja2==3.1.3